+++ controllers/factory.go
@@ -82,6 +82,11 @@ func (f *Factory) Build(ctx context.Context) (*Reconcilers, error) {
 	return &f.reconcilers, nil
 }
 
+// Close cleans up any open resources from the created dependencies.
+func (f *Factory) Close(ctx context.Context) error {
+	return f.deps.Close(ctx)
+}
+
 func (f *Factory) WithClusterReconciler(capiProviders []clusterctlv1.Provider) *Factory {
 	f.dependencyFactory.WithGovc()
 	f.withTracker().WithProviderClusterReconcilerRegistry(capiProviders).withAWSIamConfigReconciler()
@@ -308,7 +313,10 @@ func (f *Factory) withTinkerbellClusterReconciler() *Factory {
 			return nil
 		}
 
-		f.tinkerbellClusterReconciler = tinkerbellreconciler.New()
+		f.tinkerbellClusterReconciler = tinkerbellreconciler.New(
+			f.manager.GetClient(),
+			f.ipValidator,
+		)
 		f.registryBuilder.Add(anywherev1.TinkerbellDatacenterKind, f.tinkerbellClusterReconciler)
 
 		return nil