+++ pkg/api/v1alpha1/vsphereclonemode_types.go
@@ -1,0 +1,19 @@
+package v1alpha1
+
+// +kubebuilder:validation:Enum=fullClone;linkedClone
+
+// CloneMode describes the clone mode to be used when cloning vSphere VMs.
+type CloneMode string
+
+const (
+	// FullClone indicates a VM will have no relationship to the source of the
+	// clone operation once the operation is complete. This is the safest clone
+	// mode, but it is not the fastest.
+	FullClone CloneMode = "fullClone"
+
+	// LinkedClone means resulting VMs will be dependent upon the snapshot of
+	// the source VM/template from which the VM was cloned. This is the fastest
+	// clone mode, but it also prevents expanding a VMs disk beyond the size of
+	// the source VM/template.
+	LinkedClone CloneMode = "linkedClone"
+)