+++ pkg/providers/tinkerbell/stack/stack_test.go
@@ -235,6 +235,7 @@ func TestTinkerbellStackInstallWithDifferentOptions(t *testing.T) {
 					"-e", gomock.Any(),
 					"-e", gomock.Any(),
 					"-e", gomock.Any(),
+					"-e", gomock.Any(),
 				)
 			}