+++ test/e2e/oidc.go
@@ -1,0 +1,39 @@
+//go:build e2e
+// +build e2e
+
+package e2e
+
+import (
+	"github.com/aws/eks-anywhere/pkg/api/v1alpha1"
+	"github.com/aws/eks-anywhere/test/framework"
+)
+
+func runOIDCFlow(test *framework.ClusterE2ETest) {
+	test.GenerateClusterConfig()
+	test.CreateCluster()
+	test.ValidateOIDC()
+	test.StopIfFailed()
+	test.DeleteCluster()
+}
+
+func runTinkerbellOIDCFlow(test *framework.ClusterE2ETest) {
+	test.GenerateClusterConfig()
+	test.GenerateHardwareConfig()
+	test.PowerOffHardware()
+	test.CreateCluster(framework.WithForce(), framework.WithControlPlaneWaitTimeout("20m"))
+	test.ValidateOIDC()
+	test.StopIfFailed()
+	test.DeleteCluster()
+	test.ValidateHardwareDecommissioned()
+}
+
+func runUpgradeFlowWithOIDC(test *framework.ClusterE2ETest, updateVersion v1alpha1.KubernetesVersion, clusterOpts ...framework.ClusterE2ETestOpt) {
+	test.GenerateClusterConfig()
+	test.CreateCluster()
+	test.ValidateOIDC()
+	test.UpgradeClusterWithNewConfig(clusterOpts)
+	test.ValidateCluster(updateVersion)
+	test.ValidateOIDC()
+	test.StopIfFailed()
+	test.DeleteCluster()
+}