+++ docs/content/en/docs/reference/changelog.md
@@ -6,6 +6,17 @@ weight: 35
 
 ## Unreleased
 
+## [v0.14.0](https://github.com/aws/eks-anywhere/releases/tag/v0.14.0)
+
+### Added
+- Add support for EKS Anywhere on AWS Snow ([#1042](https://github.com/aws/eks-anywhere/issues/1042))
+- Static IP support for BottleRocket ([#4359](https://github.com/aws/eks-anywhere/pull/4359))
+- Add registry mirror support for curated packages
+- Add copy packages command ([#4420](https://github.com/aws/eks-anywhere/pull/4420))
+
+### Fixed
+- Improve management cluster name validation for workload clusters
+
 ## [v0.13.1](https://github.com/aws/eks-anywhere/releases/tag/v0.13.1)
 
 ### Added